+++ data-raw/create sysdata.R
@@ -5,16 +5,16 @@ logo <- magick::image_read_pdf("data-raw/GrattanSVGLogo.pdf")
 logogrob <- grid::rasterGrob(logo)
 
 chart_types <- tibble::tribble(
-                         ~type, ~width, ~height, ~caption, ~title,  ~subtitle, ~class,
-                      "normal",  22.16,    14.5,      120,     70,         75, "normal",
-                  "normal_169",  30.00,    14.5,      180,     95,        100, "normal",
-                        "tiny",  22.16,   11.08,      120,     70,         75, "normal",
-                 "wholecolumn",  22.16,   22.16,      120,     70,         75, "normal",
-                    "fullpage",  44.32,   22.16,      240,    140,        150, "normal",
-                   "fullslide",   25.4,   19.05,      140,     55,         70, "fullslide",
-               "fullslide_169",   33.87, 19.05,       175,     55,         95, "fullslide",
-                "fullslide_44",   25.4,  25.4,        140,     50,         95, "fullslide",
-                        "blog",   25.4,   19.05,      155,     62,         85, "fullslide")
+                         ~type, ~width, ~height, ~caption, ~title,  ~subtitle, ~class, ~pptx_template,
+                      "normal",  22.16,    14.5,      120,     70,         75, "normal", NA_character_,
+                  "normal_169",  30.00,    14.5,      180,     95,        100, "normal", NA_character_,
+                        "tiny",  22.16,   11.08,      120,     70,         75, "normal", NA_character_,
+                 "wholecolumn",  22.16,   22.16,      120,     70,         75, "normal", NA_character_,
+                    "fullpage",  44.32,   22.16,      240,    140,        150, "normal", NA_character_,
+                   "fullslide",   25.4,   19.05,      140,     55,         70, "fullslide", "template_43.pptx",
+               "fullslide_169",   33.87, 19.05,       175,     55,         95, "fullslide", "template_169.pptx",
+                "fullslide_44",   25.4,  25.4,        140,     50,         95, "fullslide", NA_character_,
+                        "blog",   25.4,   19.05,      155,     62,         85, "fullslide", NA_character_)
 
 
 blog_border <- 0.15